FEATURE: rename project to Nero Launcher for personal use
