+++ app/src/main/AndroidManifest.xml
@@ -9,11 +9,11 @@
     <uses-permission android:name="android.permission.REQUEST_DELETE_PACKAGES" />
 
     <application
-        android:name=".app.RlApplication"
+        android:name=".app.NeroLauncherAndroidApp"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.CustomLauncher"
+        android:theme="@style/Theme.NeroLauncher"
         android:usesCleartextTraffic="true">
 
         <activity
