+++ app/src/main/java/com/neronguyenvn/nerolauncher/core/common/coroutine/Dispatcher.kt
@@ -1,0 +1,9 @@
+package com.neronguyenvn.nerolauncher.core.common.coroutine
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Dispatcher(val dispatcher: ClDispatcher)
+
+enum class ClDispatcher { Default, IO }
