+++ app/src/main/java/com/neronguyenvn/nerolauncher/core/data/di/DataModule.kt
@@ -1,0 +1,20 @@
+package com.neronguyenvn.nerolauncher.core.data.di
+
+import com.neronguyenvn.nerolauncher.core.data.AppRepository
+import com.neronguyenvn.nerolauncher.core.data.repository.OfflineFirstAppRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class DataModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindsApplicationRepository(
+        appRepo: OfflineFirstAppRepository
+    ): AppRepository
+}
