+++ app/src/main/java/com/neronguyenvn/nerolauncher/core/ui/pageslider/PageIndicator.kt
@@ -1,0 +1,40 @@
+package com.neronguyenvn.nerolauncher.core.ui.pageslider
+
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PageIndicator(
+    index: Int,
+    count: Int,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = modifier
+    ) {
+        repeat(count) {
+            val size by animateDpAsState(if (index == it) 12.dp else 8.dp)
+            Spacer(
+                modifier = Modifier
+                    .padding(horizontal = 4.dp)
+                    .size(size)
+                    .background(
+                        color = MaterialTheme.colorScheme.primary,
+                        shape = RoundedCornerShape(30)
+                    )
+            )
+        }
+    }
+}
