--- app/src/main/java/com/neronguyenvn/reorderablelauncher/core/common/coroutine/Dispatcher.kt
@@ -1,9 +1,0 @@
-package com.neronguyenvn.reorderablelauncher.core.common.coroutine
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Dispatcher(val dispatcher: ClDispatcher)
-
-enum class ClDispatcher { Default, IO }
