--- app/src/main/java/com/neronguyenvn/reorderablelauncher/core/designsystem/util/DesignSystemUtils.kt
@@ -1,42 +1,0 @@
-package com.neronguyenvn.reorderablelauncher.core.designsystem.util
-
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-/*@Composable
-fun Modifier.noRippleClickable(onClick: () -> Unit): Modifier {
-    return this then clickable(
-        indication = null,
-        interactionSource = remember { MutableInteractionSource() }
-    ) {
-        onClick()
-    }
-}*/
-
-fun Drawable.asBitmap(): Bitmap {
-    if (this is BitmapDrawable) {
-        return bitmap
-    }
-
-    val bitmap = Bitmap.createBitmap(intrinsicWidth, intrinsicHeight, Bitmap.Config.ARGB_8888)
-    val canvas = Canvas(bitmap)
-    setBounds(0, 0, canvas.width, canvas.height)
-    draw(canvas)
-
-    return bitmap
-}
-
-@Composable
-inline fun Modifier.conditional(
-    condition: Boolean,
-    ifTrue: Modifier.() -> Modifier,
-    ifFalse: Modifier.() -> Modifier = { this },
-): Modifier = if (condition) {
-    then(ifTrue(this))
-} else {
-    then(ifFalse(this))
-}
