+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Reorderable Launcher</string>
+    <string name="app_name">Nero Launcher</string>
 </resources>
