+++ app/src/main/res/values/themes.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.CustomLauncher" parent="android:Theme.Material.Light.NoActionBar">
+    <style name="Theme.NeroLauncher" parent="android:Theme.Material.Light.NoActionBar">
         <item name="android:windowShowWallpaper">true</item>
         <item name="android:windowBackground">@android:color/transparent</item>
     </style>
